backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.1
python-multipart>=0.0.19
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.58.1
pypdf>=5.1.0
python-pptx>=1.0.2
reportlab>=4.2.5
python-docx>=1.2.0
pillow>=11.0.0
httpx>=0.28.1
python-dotenv>=1.0.1
tiktoken>=0.8.0
numpy>=1.26.0
razorpay>=1.4.2
